requirements.txt
# Flask and web framework
flask==3.0.0
werkzeug==3.0.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Configuration and environment
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Encryption
cryptography==41.0.7

# Google Gemini API (New SDK - GA as of May 2025)
google-genai==0.2.2

# HTTP requests
requests==2.31.0
urllib3<2.0.0  # Pin to v1.x for compatibility with older OpenSSL/LibreSSL versions

# Sitemap parsing
ultimate-sitemap-parser==0.5

# Apify integration
apify-client==1.7.1

# Production server
gunicorn==21.2.0

# Optional: async support
# aiohttp==3.9.1
# asyncio==3.4.3
